Fix r124/ydb197 subtest failure due to reference file issue
ydb_dist and gtm_dist env vars get set to their real path by ydb_env_set.
That is, if these env vars point to a directory D1 which is a soft link to another directory D2,
these env vars would be set to D2 by ydb_env_set. SOURCE_REALPATH is the tag to use in the
reference file for this purpose. It was already used for the UTF-8 part of the reference file
but was missed out for the non-UTF8 part. Now fixed.