Enhance go/simpleapithreeenp1f subtest to test https://gitlab.com/YottaDB/Lang/YDBGo/merge_requests/62
https://gitlab.com/YottaDB/Lang/YDBGo/merge_requests/62 in YDBGo repo fixes an issue in the GoWrapper
where if more than one line of input was provided to the go code that implemented the 3n+1 algorithm,
the second line of input would have incorrect output show up. To test that without lengthening the
test run time any more than it already is, the 250,000 input is split into 150,000 and 100,000 and
the reference file is updated to have 2 lines of output. Without the fixes to the mentioned merge
request, this test fails and passes with the fixes.