Enhance unicode_io/unicode_fifo subtest even in jobbed off processes in the subtest to better analyze rare failures
A prior commit (SHA e553a88156344257feeae7c020039cef4f79a89a) added "do ^sstepgbl" in the parent M program
but left it out in a couple of places where we jobbed off an M process. This caused incomplete state
information (^sstepgbl recording was happening in only the parent M program, not the child).
This commit ensures the child also does ^sstepgbl recording that way giving us the complete timeline of
events in case of a rare test failure due to some timing issue in the test.