Fix rare r120/ydbdist subtest failure due to extra SECSHRNOYDBDIST messages from concurrently running tests
As part of a prior commit (SHA 386a38c8898718c03f1ca08e18305ade5aa6ac68), this failure was supposedly
fixed but it was an incomplete fix. The file syslog1.txt was created by a call to getoper.csh
with the search pattern which had the desired pid followed by a SECSHRNOYDBDIST message in it.
But what was missed out in that commit was that the file syslog1.txt would contain the desired message
and all messages before that from the syslog. It is possible there are SECSHRNOYDBDIST messages in
the syslog from unrelated pids before the desired SECSHRNOYDBDIST message in syslog1.txt. Therefore
the correct fix is to also search for the specific pid in the following check_error_exist.csh call.
This is now done.