Fix rare expect/gtm4661 subtest failure with TEST-E-LSOF error due to mumps processes still accessing the database; Enhance test framework to capture ps listing at time of TEST-E-LSOF error
The expect/gtm4661 subtest occasionally failed with a TEST-E-LSOF error.
This was because a "mumps -run waitforjobtodie^sigproc" process started in gtm4661b.exp was still
running when the dbcheck.csh happened. This is because we issue this command in gtm4661b.exp but
do not wait for the shell prompt to show up before sending the "exit" command. Therefore, the
expect invocation of gtm4661b.exp returns even while the "mumps" command is still running.

To fix this, the "mumps -run waitforjobtodie" invocation is moved to the parent gtm4661.csh script
where we don't have the complications of dealing with an expect terminal (send and expect commands).

While at this, the test is reworked to use expect.dbg and expect_sanitized stuff just like other
tests are. This will help us better debug any rare test failures with the detail of exactly what
happened.

And the test framework (com/errors.csh) has been enhanced to also capture a ps -ef --forest at the
time when it generates the TEST-E-LSOF error. This is needed because the lsof_* report only contains
pids whereas what we want is the know what that pid was which the ps -ef listing gives us. This is
what helped identify the leftover process as the "mumps -run waitforjobtodie^sigproc" process.