Fix triggers/gtm7509 subtest failures due to random sprgde files created by test framework for a multi-region dbcreate interfering with a single-region dbcreate in the same test
We had a test failure where the first dbcreate done in gtm7509.csh (it does 5 dbcreates) was for
a 1-region database and failed with the following error.

%GDE-E-OBJNOTFND, Region AREG does not exist

This is because it was using a .sprgde file (created by the test framework from a prior test run
to test the spanning regions functionality) corresponding to a multi-region dbcreate done in a
later stage in the same test across even the single-region dbcreate calls.

To avoid this issue, one needs to define the env var gtm_test_sprgde_id to different values for each
dbcreate done in the test that way different .sprgde files get created for different dbcreate calls
and one does not interfere with the other.