Fix occasional simplethreadapi/fatalerror1 and simplethreadapi/fatalerror2 subtest failures on RHEL7 due to FATALERROR1/FATALERROR2 messages not showing up in syslog
The simplethreadapi/fatalerror1 subtest failed on a RHEL7 box because a FATALERROR1 message
did not show up in the syslog as expected. This is a case where a send_msg() of FATALERROR1
was done in ydb_simpleapi_ch() but it showed up in syslog as the following instead.

Feb 09 10:12:43 hostname fatalerror[78806]: out of memory [78806]
Feb 09 10:12:43 hostname fatalerror[78806]: out of memory [78806]

The same issue happens in the fatalerror1 and fatalerror2 subtests. And interestingly only in
the simplethreadapi test, not the simpleapi test which also has almost the same subtests.
So something to do with having multiple threads in the same process. And this happens only on
a RHEL7 x86_64 box in our test cluster.

When I ran the test without replay for 20 runs, I got 5 failures so the failure rate is 25%.
I could not find a pattern in the random setting choice of the test system framework either.
So I suspect it is some issue with syslog() call which notices no more memory only in some cases
and decides to send out a standard template message instead of the desired FATALERROR1 or
FATALERROR2 message.

I have fixed the test to allow for "fatalerror.*out of memory" messages too in addition to the
FATALERROR1/FATALERROR2 message.