Fix go test csh script issue causing failure status to be incorrectly reported; Enhance test to log <go get> steps and timing to better debug net/http TLS handshake timeouts
The go/threeenp1B1 subtest failed with the following diff on an ARMV6L box in internal testing.

 # Running : go get -t lang.yottadb.com/go/yottadb
 package lang.yottadb.com/go/yottadb: unrecognized import path "lang.yottadb.com/go/yottadb" (https fetch: Get https://lang.yottadb.com/go/yottadb?go-get=1: net/http: TLS handshake timeout)
 TEST-E-FAILED : [go get -t lang.yottadb.com/go/yottadb] returned failure status of 0
 [source go/u_inref/setupgoenv.csh] failed with status [0]. Exiting...

The "go get" command clearly failed but the failure status shows up as 0 in the last 2 lines above.
The issue is the way $status is used in the setupgoenv.csh script.

$status is a special variable that is set after each statement by tcsh to the exit status of that
statement. So if $status is checked in say an "if" statement, the same value can no longer be
accessed as $status in later statements inside the "if" block. The $status usage there
would reflect the success of the "if" check (which would be 0) whereas what is intended is the
same non-zero $status value found during the "if" check.

Therefore, the fix is to set the exit status of "source setupgoenv.csh" in a different
variable say "$status1" and use this in the "if" check as well as inside the "if" block
without any issues.

In addition the <go get> command is now passed more qualifiers to record each step as it happens and
the time of each step is also recorded to better debug the net/http timeout failures.