Fix rare rundown_argless/ipcsnoreset subtest failure due to argumentless mupip rundown incorrectly removing db semaphore before imptp process has finished db rundown
This subtest failed an assert in one test run (out of dozens of test runs).

 #0  __pthread_kill () at ../sysdeps/unix/sysv/linux/pthread_kill.c:56
 #1  gtm_dump_core () at sr_unix/gtm_dump_core.c:72
 #2  ch_cond_core () at sr_unix/ch_cond_core.c:76
 #3  rts_error_va () at sr_unix/rts_error.c:194
 #4  rts_error_csa () at sr_unix/rts_error.c:101
 #5  gds_rundown () at sr_unix/gds_rundown.c:360
 #6  gv_rundown () at sr_port/gv_rundown.c:123
 #7  gtm_exit_handler () at sr_unix/gtm_exit_handler.c:243
 #8  __run_exit_handlers () at exit.c:108
 #9  __GI_exit () at exit.c:139
 #10 gtm_image_exit (status=-15) at sr_unix/gtm_image_exit.c:27
 #11 deferred_exit_handler () at sr_unix/deferred_exit_handler.c:122
 #12 deferred_signal_handler () at sr_port/deferred_signal_handler.c:56
 #13 t_end () at sr_port/t_end.c:1790
 #14 gvcst_put2 () at sr_port/gvcst_put.c:2641
 #15 gvcst_put () at sr_port/gvcst_put.c:299
 #16 op_gvput () at sr_port/op_gvput.c:77
 #17 ydb_set_s () at sr_unix/ydb_set_s.c:131
 #18 ydb_stm_threadq_dispatch () at sr_unix/ydb_stm_thread.c:385
 #19 ydb_stm_threadq_process () at sr_unix/ydb_stm_thread.c:262
 #20 ydb_stm_thread () at sr_unix/ydb_stm_thread.c:99
 #21 start_thread () at pthread_create.c:486
 #22 clone () at ../sysdeps/unix/sysv/linux/x86_64/clone.S:95

gds_rundown.c
-------------
    357   /* Check # of processes counted on access sem. */
    358   if (-1 == (semval = semctl(udi->semid, DB_COUNTER_SEM, GETVAL)))
    359   {
--> 360           assert(FALSE);
    361           rts_error_csa(CSA_ARG(csa) VARLSTCNT(12) ERR_CRITSEMFAIL, 2, DB_LEN_STR(reg), ERR_SYSCALL, 5,
    362                         RTS_ERROR_TEXT("gds_rundown SEMCTL failed to get semval"), CALLFROM, errno);
    363   }

(gdb) f 5
 #5  gds_rundown () at sr_unix/gds_rundown.c:360
360                                     assert(FALSE);

(gdb) p udi->semid
$2 = 1912799236

(gdb) p save_errno
$3 = 22			/* EINVAL implying semid does not exist, i.e. it was already removed */

And in the syslog, one can see this semid was indeed removed by an argumentless rundown.

YDB-MUPIP-INSTANCE1[4845]: %YDB-I-MURNDWNARGLESS, Argumentless MUPIP RUNDOWN started ... from directory rundown_argless_1/ipcsnoreset/
YDB-MUPIP-INSTANCE1[4845]: %YDB-I-SEMREMOVED, Semaphore id 1912799236 removed from the system

The argumentless rundown decided to remove this semaphore because it found it orphaned even though
all imptp processes attached to the database had not yet done the rundown. This is possible in the
following scenario.

Process P1 bumps the semaphore counter to 16K (because this test had set the gtm_db_counter_sem_incr
env var to 16K). Process P2 bumps the semaphore counter to 32K. Process P3, P4, ... P10 (the
remaining 8 imptp processes) do not bump the semaphore counter because the counter is already the
max of 32K. Now if P1 and P2 finish db rundown before at least one of P3 thru P10 finish their db
rundown, the process exit of P1 and P2 will cause the semaphore counter value to be undone (we use
SEM_UNDO) so the semaphore counter would go back to 0, so the semaphore counter would be 0 even
though one of P3 thru P10 is still attached to the database. An argumentless MUPIP RUNDOWN has no
way of knowing this and so will go ahead and consider the sem counter of 0 as an orphaned one and
remove it thereby causing this test failure.

This scenario is not possible in real life where each process bumps the counter only by 1. We need
32K processes to bump the counter by 1 each and more processes to attach to the db but not bump
the counter and then all the first 32K processes to rundown the db but the later process to still
attach and a mupip rundown to be attempted. This is almost impossible in practice so this is not
a user-visible issue.

To address the test failure, disabling the artificial semaphore counter overflow this test creates
is enough. That prevents the semaphore from being incorrectly treated as orphaned by the mupip
rundown. And lets the script test the original purpose of writing this test (to test mupip rundown
and imptp process rundown happening concurrently).