Fix rare timing/gctest subtest failure due to one iteration taking more than 3x time the previous iteration
We occasionally see failures in this subtest with the following symptom.

TEST-E-FAIL : cpu(2)=0.095 is 3.39x times cpu(1)=0.028 but max allowed is 3x
cpu(1)="0.028"
cpu(2)="0.095"
cpu(3)="0.237"
cpu(4)="0.485"

In this case, cpu(2) is slightly more than 3x times cpu(1).  But cpu(3) is a lot less than 3x times
cpu(2). And cpu(4) is a lot less than 3x times cpu(3).  Therefore cpu(4) is still 9x times less
than cpu(1). Which effectively tests that objective of the test that each iteration does not go
more than 3x the previous iteration on average.

Therefore the above failure is a false alarm. The test is now fixed to signal a failure
only if cpu(4) is 7.5x times more than cpu(1). This is a lot more stricter than before
(since previously a test could pass as long as cpu(4) was less than 27x times cpu(1)).

With these changes, the occasional failures we see no longer happen.