Enhance triggers/gtm7522a subtest to cause test failure if ever any $ZTRIGGER command fails
In case $ZTRIGGER commands fail (for example, due to a bug), it is still possible for the
gtm7522a subtest to silently pass. This is because we do not check the return value of $ZTRIGGER.
The M program is now enhanced to check the return value and signal a "TEST-E-FAIL" in that case.