Fix random longname/objtest and r100/dllversion subtest failures
Currently R124 is the latest YottaDB release. The GT.M version that this release is based on is V63005.
Recently GT.M V63006 and V63007 were released on Sourceforge and so those were built in-house.
But this posed a problem in a few tests that use get_max_ver_dynamically.csh. This is because if we
run say the longname/objtest subtest with a V63005_R124 build, the call to get_max_ver_dynamically.csh
would notice V63006/V63007 builds and would see their object file format is different from the R124
object file format and so incorrectly determine the maximum version to be V63007. This meant that
com/random_ver.csh could later choose V63005_R124 as the random version which would be incorrect since
that would most likely have the same object file format as the currently tested build of the YottaDB
mainline. This in turn would cause the subtest to fail since it assumes that the random version chosen
is guaranteed to have a different object file format.

The fix is to check if the current production version (gtm_curpro env var) is defined (if defined,
it would be V63005_R124 until R126 is released) and if so ensure the list of versions never gets
populated to more than V63005_R124. If this env var is not defined, then this list would never go
past the current version used for testing (like before).