Fix random longname/objtest and r100/dllversion subtest failures for real
A prior fix for this (commit SHA fc29910e327c56b9892f2d4833cd853509179ea8) was an incomplete
change. This is because it relied on the assumption that gtm_curpro env var would be inherited
by the test framework from the calling environment. Turns out this was not the case not just
for this env var but a lot of other env vars that were being invoked using "check_setenv_echo"
in gtm_env_missing.csh. The issue was in an awk file which is supposed to check if the
env var exists using $?<envvarname> but instead was doing just $? resulting in a completely
different check. So instead of checking "if ($?gtm_curpro == 0)", we ended up checking "if ($? == 0)"
which meant gtm_curpro was always considered as undefined and set to the current running test
version instead of being inherited. This is now fixed by fixing the "check_setenv_echo" macro to
work just like the "check_setenv" macro in com/process_defaults.awk.