Randomly enable Go race detector in all subtests in the go test (enabled if env var ydb_go_race_detector_on set set to non-zero value)
https://gitlab.com/YottaDB/DB/YDBTest/merge_requests/672 fixes an issue in a test system Go program
which resulted in a Go garbage collector panic occasionally. The output used to be of the form

> runtime: nelems=32 nalloc=5 previous allocCount=4 nfreed=65535
> fatal error: sweep increased allocation count

Interestingly, this got fixed once a race condition, that was alerted by the Go race detector,
in the test program was fixed.

Given that, the go test suite is enhanced to run all subtests with "go build" or "go test" randomly
having "-race" added so the race detector is enabled.