Fix simplethreadapi/utilfuncs subtest failure (assert(fork_after_ydb_init) failure in ydb_stm_atfork_child() due to multiple threads doing fork() at same time)
In a YottaDB process, fork() can be done in multiple threads as long as the application ensures
only one fork() executes at any point in time. This was not being ensured in utilfuncs3_MT_STAPIFORKEXEC.c
and is now fixed by moving the mutex lock to be a global variable instead of a thread-specific variable
and initializing it globally before spawning off any threads.