[YottaDB/Lang/YDBGo#8] Fix framework script setupgoenv.csh to not try build/install as part of go get
Recent changes as part of YottaDB/Lang/YDBGo#8 made the "master" branch of YDB incompatible with
the "master" branch of YDBGo. go get without the "-d" qualifier tries to do a build of the "master"
YDBGo branch using the "master" YDB branch and fails with the following error.

go/src/lang.yottadb.com/go/yottadb/variadic.go:67:13: could not determine kind of name for C.ydb_call_variadic_plist_func_st

But the "develop" branch of YDBGo is compatible with the "master" YDB branch.
The test anyways switches to the "develop" branch soon afterwards.
So best to let the build happen after that switch (as part of "go build" or "go test") and skip
the build as part of the "go get".

The "-d" qualifier as part of "go get" does exactly that.