Fix unicode_io/unicode_fifo subtest failure due to GBLOFLOW error on 32-bit platforms with MM access method
On the 32-bit ARMV7L or ARMV6L boxes, we occasionally see this subtest fail with a GBLOFLOW error.
That is because the ^sstepgbl global (which records every M line as it gets executed in the global)
becomes too huge and the database exceeds 1Gb in size. That is too much for 32-bit boxes where the
database is mmaped (i.e. MM access method) whose virtual address space is only 4Gb.

The ^sstepgbl scheme was originally introduced to debug timing issues but it is now causing test
failures. So this scheme is removed in case the platform is 32-bit and the current access method
choice by the test framework is MM. If we later encounter test timing issues in this configuration,
we have to debug those using a different approach than ^sstepgbl.