[V63006] Fix longname/rout_label subtest failure now that do ^rtnname when rtnname is longer than 31 chars looks for a .o file with name truncated to 31 bytes (GTM-8178)
And because of the nature of the flow of the test such a file does exist.
And therefore, one does not get a ZLINKFILE/FILENOTFND error like before from a DO ^RTNNAME
whereas previously (GT.M V6.3-005 and before) one needed to do an explicit ZLINK of the long name.