[V63006] Fix io/gtm8239 subtest failure now that FILENOTFND error prints absolute path name which exceeds terminal width causing $y to be 1 more than before
case6 is where we noticed $y become 4 instead of the expected 3.
This is because the FILENOTFND message got split across 2 lines each of width 80.
Previously only the relative path of the M file was printed which caused the message
to be within 80. Now it is not enough.

The fix is to source the test framework expectinit.exp script which sets the terminal
column limit to 132 and that is good enough for even the new FILENOTFND message with
absolute file path.