[V63006] Fix relink/zrupdate and relink/setzroutines subtest failures now that ZCOMPILE without a .m extension automatically adds the .m extension (GTM-8178)
ZCOMPILE bar2.edit worked previously where the input file name bar2.edit was taken as is
but starting V63006 it assumed the input file name as bar2.edit.m (added the .m extension
automatically) whereas the file that existed in the test was bar2.edit. This caused a
FILENOTFND error from the ZCOMPILE. Therefore the test scripts are enhanced to add a .m
extension at the end of already existing extensions.