[V63006] Fix occasional dse/dse_all subtest failure with -replic due to source server failing with DBFLCORRP error
After integrating GT.M V6.3-006, we started seeing the dse/dse_all subtest fail occasionally
with a DBFLCORRP error in the source server's log file.

Below is the C-stack of the source server when it issues the DBFLCORRP error.

 #5  rts_error_csa () at sr_unix/rts_error.c:99
 #6  grab_crit () at sr_unix/grab_crit.c:163
 #7  repl_ctl_create () at sr_port/gtmsource_ctl_init.c:181
 #8  gtmsource_ctl_init () at sr_port/gtmsource_ctl_init.c:436
 #9  gtmsource_get_jnlrecs () at sr_unix/gtmsource_process_ops.c:976
 #10 gtmsource_process () at sr_unix/gtmsource_process.c:1544
 #11 gtmsource () at sr_unix/gtmsource.c:528
 #12 mupip_main () at sr_unix/mupip_main.c:124
 #13 dlopen_libyottadb () at sr_unix/dlopen_libyottadb.c:148
 #14 main () at sr_unix/mupip.c:19

Interestingly, the test already accounts for the possibility of the DBFLCORRP error by waiting
for the source server to have sent at least one sequence number (that way it would have opened
the journal files). But the way it does this is by checking that the receiver log has received
at least one history record. The two are not the same. The history record would be sent by the
source server even without opening the journal files. Therefore the check is not correct. It is
now fixed to do a dummy update and wait for this to be sent across before going ahead with the
dse commands that manipulate the file-corrupt field in the db file header.