[DEBIAN10] Fix tls/errors and socket/mwebserver subtest failures due to <SSL_use_certificate:ee key too small> error in OpenSSL 1.1.1
The tls/errors subtest failed with the following error in various stages of the test.

> %YDB-I-TEXT, Failed to add certificate tls_1/errors/demoCA/server.pem. Reason: error:140C618F:SSL routines:SSL_use_certificate:ee key too small

This is because we use a keysize of 1024 to generate the certificates which OpenSSL 1.1.1 (on Debian 10)
refuses to accept due to it not being secure enough. The fix is to generate certificates with a keysize
of 4096.

Note that on the Debian 10 system, the default security level is 2

> cat /etc/ssl/openssl.cnf
CipherString = DEFAULT@SECLEVEL=2

whereas previously it used to be 1. Hence the stricter security requirements.
Another solution would have been to change the above from SECLEVEL=2 to SECLEVEL=1 but
that would be a less secure solution to the problem.