[YDBTest#308] Fix rare v60001/gtm7461 subtest failure due to db ipcs being left around
The v60001/gtm7461 subtest failed rarely with the following diff

12,13d11
< %YDB-E-REPLINSTFROZEN, Instance INSTANCE1 is now Frozen
< %YDB-I-REPLINSTFREEZECOMMENT, Freeze Comment: PID xxxx encountered JNLFILEOPNERR; Instance frozen
16a15,20
> v60001_0_22/gtm7461/rundown.out
> %YDB-E-MUUSERLBK, Abnormal shutdown of replication-enabled database v60001_0_22/gtm7461/a.dat detected

This is because after the endtp.csh call in gtm7461.csh, the db ipcs were still lying around.
And that is because if the last process to be attached to the database turns out to have chosen
M call-in to impjob^imptp for this process, it did an os.Exit() after the call-in instead of
just a "return". This meant that the deferred call to yottadb.Exit() done at the beginning of
main() in impjobgo.go does not get run in this case and not running the YottaDB exit handler
causes the db ipcs to be left around. Fixing the os.Exit() to instead be a "return" ensures
the YottaDB exit handler does get run and in turn addresses the rare test failure.