Fix various subtest failures in <go> test due to warning from <go build> or <go test>
After a system patch on an in-house Arch Linux box, "go build" started issuing a warning message.

/usr/sbin/ld: warning: libyottadb.so: corrupt GNU_PROPERTY_TYPE (5) size: 0

Not sure why this shows up. Likely some regression in either the system ld or the go packages.
Interestingly the "go build" succeeded (i.e. exit status of 0) except for this warning.
Therefore the test is fixed to redirect the output of "go build" and "go test" to log files
and just check the exit status. This lets the test pass without issues.

For the record, below is the version of go and ld on the Arch Linux box at the time of the failure.

go version go1.12.4 linux/amd64
GNU ld (GNU Binutils) 2.32