Disable r124/jnlunxpcterr subtest (maybe temporarily) on Ubuntu/Debian/Arch Linux for x86_64 platform only as it causes CPU soft lockup
A previous commit (SHA cad3c2558a33988ee0ccb334135ed16b5e78203b) had disabled this subtest on
Ubuntu and Arch Linux but for all architectures (i.e. x86_64 as well as armv7l, armv6l etc.).
But we have seen the CPU soft lockup issue (which is a Linux kernel issue) only on x86_64
architecture so disable this subtest on Ubuntu/Arch and Debian (now that it is supported on x86_64)
but only on x86_64.

Therefore existing flags UBUNTU_LINUX etc. have been renamed as UBUNTU_LINUX_X86_64 etc.
This was possible since these flags were not used anywhere else in the test system.
And these flags are now set only if the `uname -m` output is `x86_64`.