Attempt to fix obscure test failures on ARM platform by limiting gdb stack traces of each thread to 100 frames (failure symptoms are INVOBJFILE etc.)
The burst_load/onlnbkup_src_a subtest failed on an AARCH64 box in in-house testing with the following diff.

34a35,36
> %YDB-E-INVOBJFILE, Cannot ZLINK object file pro/_XCMD.o due to unexpected format
> %YDB-E-INVOBJFILE, Cannot ZLINK object file pro/_XCMD.o due to unexpected format
36a39,41
> thorin:burst_load_1/onlnbkup_src_a/online_back_out.out
> %YDB-E-INVOBJFILE, Cannot ZLINK object file pro/_XCMD.o due to unexpected format
> %YDB-E-INVOBJFILE, Cannot ZLINK object file pro/_XCMD.o due to unexpected format

Looking at the test output showed a MUTEXLCKALERT situation during the test when the online backup
command was started which caused the C-stack to be obtained using gdb of the running impjobgo.go
process. The "thread apply all bt" command ran into a loop in gdb for reasons not yet clear. It
kept printing the same frame indefinitely before the test timed out and killed the gdb process.
We suspect this has something to do with the INVOBJFILE error above. And so, the gdb looping
is now prevented by limiting the # of frames to go back to in each thread to 100.