Fix rare r124/ydb312_gtm8182a and r124/ydb312_gtm8182g subtest failure due to ipc ids which have a search ipc id as a substring
We had a failure of the r124/ydb312_gtm8182g subtest with the following diff.

19,20c19,57
<      Number of processes attached to journal pool of mumps.repl: 1
< ##DIFFERENT EXPRESSION##^     JNLPOOL ftok : mumps.repl : \(semval=1, semncnt=0, semzcnt=0, sempid=[0-9]*\)$## VS ##dest##
---
>      Number of processes attached to journal pool of mumps.repl: dest
> dest
> dest
> dest
.
.

This was because shm1.out contained a shmid of 45. And there were various other ipcs in the system
(which show up in an `ipcs -a` output) that had 45 as a substring.

> ipcs -a | grep 45
0x00000000 458752     user       600        524288     2          dest
0x00000000 1909784585 user       777        8294400    1          dest
0x00000000 395837453  user       777        8294400    1          dest
.
.

This meant a grep of 45 found a lot of ipcs (instead of just the one we wanted). And in turn caused
the test failure.

This is fixed by doing a `grep -w` instead which will do a full-word search of 45 and identify
just the one ipc we are interested in from the `ipcs -a` output.