Fix various test failures with "User signal 1" failure symptom if v55000/GTM6813 subtest hangs due to other issues
We had the v55000/GTM6813 subtest hang on 32-bit ARMV7L boxes after integrating GT.M V6.3-006.
That is an issue which will be separately investigated.

But once the hung test timed out (test framework has a timeout) various subsequent tests failed
with a "User signal 1" symptom because the GTM6813 subtest had a jobbed off process (say INTERRUPT)
which kept sending MUPIP INTRPTs to a specific pid (say MAINPID) forever (until MAINPID set a global
^quit).  Eventually (in a day or two) MAINPID errored out with a MEMORY error but the INTERRUPT pid
still kept running and was sending SIGUSR1 (which is SIG-10 on Linux) to MAINPID so any other running
test which got the same pid as MAINPIND failed with a "User signal 1" symptom due to an interrupt
at a point where the subsequent test was not expecting it.

This commit addresses the secondary failure.