Fix rare simpleapi/locks and simplethreadapi/locks subtest failure in case Lock timeout test takes 2 seconds instead of 1 second
We had the simplethreadapi/locks subtest fail with the following diff on one ARMV6L box.

42c42
< Lock timeout test PASSED for ydb_lock_st() : Timeout expected = 1 second. Actual timeout = 1 seconds
---
> Lock timeout test PASSED for ydb_lock_st() : Timeout expected = 1 second. Actual timeout = 2 seconds

If the ydb_lock_s() or ydb_lock_st() times out in 1 or 2 seconds, the test treats that as a PASS
situation but the output printed is the actual timeout which could be 1 or 2 seconds. The reference
file though accepts only 1 second even though the test C code accepts 1 or 2 seconds as PASS.
The test is fixed to have 1 or 2 seconds as acceptable even in the reference file.