Fix rare v60000/gtm4525b subtest failure due to MUINSTFROZEN message taking a little more than 15 seconds to show up
On an ARMV7L box we had a test failure with the following diff.

105a106,108
> TEST-E-NOTFOUND message MUINSTFROZEN not found in set_frozen.out after 15 secs
> Waited from: Sat May 11 09:58:41 EDT 2019
> to: Sat May 11 09:58:58 EDT 2019

The script v60000/u_inref/gtm4525b.csh had issued the following commands.

($MUPIP set -region "*" -journal=before -replic=on >&! set_frozen.out & ; echo $! >! mupip_set_pid.out)	>> mupip_set_bkgrnded.out
$gtm_tst/com/wait_for_log.csh -message MUINSTFROZEN -log set_frozen.out -duration 15

And in the test output below is the timestmap of the corresponding files.

-rw-rw-r--    6 2019-05-11 09:58:41.000000000 -0400 mupip_set_pid.out
-rw-rw-r--   10 2019-05-11 09:58:41.000000000 -0400 mupip_set_bkgrnded.out
-rw-rw-r-- 2504 2019-05-11 09:59:03.000000000 -0400 set_frozen.out

The file set_frozen.out does have the MUINSTFROZEN message but it showed up just 1 second after when
the wait stopped at 09:58:58. Therefore this is a case of the 15 second timeout not being enough.
On a loaded system, it is possible the MUPIP SET command takes more than 15 seconds to run.
Therefore the timeout is now bumped to be 60 seconds instead.