Fix rare io/pipetest subtest hang
An in-house test run of the io/pipetest subtest hung (actually was in a spin-loop) at line 75 below.

io/inref/fix.m
--------------
     63    write !,"This test has a recordsize of 60, but is not fixed format",!
     64    write "Writing 10 lines each with each one containing 63 characters",!
     65    write " and expect to output 60 chars on one line followed by a newline then by 3 on the next line",!,!
     66    ;we have to do 3 reads for each write or the pipe will fill up with more iterations
     67    set p="test"
     68    open p:(comm="cat -u":recordsize=60)::"pipe"
     69    zshow "d"
     70    for i=1:1:10 do
     71    . use p
     72    . write "123456789012345678901234567890123456789012345678901234567890abc",!
     73    . for  read x(i):0 quit:x(i)'=""
     74    . read y(i)
     75    . for  read z(i):0 quit:z(i)'=""

The test does a write of 63 characters at line 72 and expects to read the first 60 into x(i) at line 73
and the empty string into y(i) at line 74 and the last 3 characters into z(i) at line 75. This is because
the recordsize is set at 60 which means the 63 character write will be split into a 60 character write
followed by a newline and then followed by 3 characters followed by a newline.

But in the hung test, x(i) contained only 57 characters and y(i) contained the 6 characters "890abc"
which meant z(i) contained the empty string after a read since there is nothing else to read. Since
line 75 has a for loop, the read z(i) was being reattempted indefinitely resulting in a spinloop.

The reason x(i) read only 57 characters is because the read had a :0 (i.e. zero timeout) specified.
This means it will attempt the read only once and return even if it did not see a newline or end of
the 60-byte record. This meant it was possible for x(i) to read a PART of the write done at line
72 (since we are not guaranteed all the 60 characters will be written to the pipe in one system
call atomically).  And that in turn meant when the remaining 6 characters are written in the pipe
device by "cat" y(i) would read those 6 characters disrupting the flow of the test.

To fix this, the for loop of zero-timeout reads is replaced by a single read with no timeout.
Not sure why the for loop was there in the first place.

With this change, the test reliably passes whereas it used to fail occasionally on one x86_64 box.