Fix occasional v52000/D9G12002636 subtest failure on loaded system
Once in a while we see a failure of the v52000/D9G12002636 subtest with the following diff.

41c41
< ..
---
> ...

The reference file expects 2 dots but we got 3 dots instead. After the first dot, there is a dot
for each second we were waiting for the interrupter process to shutdown. So in the failure case,
the interrupter process took 2 seconds to shutdown instead of the expected 1 second. On a loaded
or slow system (like the ARM) it is possible for the shutdown process to take more than 1 second
to shut down. To avoid false test failures in that case, the entire block of code which prints
a dot for each second of wait is removed.