Fix go test failures due to TLS handshake timeout errors from go get on 1-CPU ARM boxes
A previous commit (SHA 3668279d8b05ef88e5b36b13286333fb3d8cc781) attempted a fix but there was a
typo. grep returns 0 in case of a match found.  Therefore a check for "! $status" was instead coded
as "$status" negating that commit completely.  This is now fixed to have the "! " prefix.