Correctly generate gdb C-stack traces even for core file generated in a subdirectory (e.g. in go/pseudoBank subtest)
Use $core (we already are in a for loop processing one core file at a time) instead of core*
which is a typo and could match more than one core file (if there are multiple core files in the
current directory) OR could match no core file (if all core files are in subdirectories).

We had one failure of the go/pseudoBank subtest where the core was in a subdirectory and we got
a garbage C-stack because the corresponding executable was not correctly determined.