[YDB#446] Fix call_ins/gtmxc_test_types and call_ins/xc_test_types subtest failures due to return value ydb_string_t length not set correctly before gtm_ci() call
After the YDB#446 code changes, the above two subtests failed with the following diff

> 150375522,(Call-In),%YDB-E-INVSTRLEN, Invalid string length 36: max 35

The issue is that in the corresponding C programs, retval8.length was not being set at all before
the gtm_ci() call. This was okay before YDB#446 code changes since the incoming ydb_string_t *
parameter length was never checked.

Now it is checked and so we need to set it to the allocated (i.e. maximum available) length before
calling gtm_ci(). This avoids a INVSTRLEN error which would otherwise show up when retval8.length
is set by one call to gtm_ci() and is used for the next call to gtm_ci() where the return value is
one byte more in length.