[V63007] Fix various subtest failures due to GTM-9073 (maximum reserved byte and REGIS/SEGIS message changes)
This involved one or more of the following changes.

* Removing some REGIS/SEGIS message occurrences as they are no longer issued by GDE in some cases.
* Adjusting the test script to work with the new reserved bytes calculation (GTM-9073).
* Adjusting the reference files to reflect the new reserved bytes calculation (GTM-9073).

Below is the release note for GTM-9073

GDE and MUPIP maintain a consistent relationship between Maximum Key Size and Maximum Reserved
Bytes. It adheres to the equation,
	Maximum Reserved Bytes = Block Size - Key Size
		- (Size of Block Header + Size of Record Header + Size of Block id + B-star Record Size),
which is equivalent to Block Size - Key Size - 32. Previously, they used inconsistent calculations
and allowed inconsistent Maximum Key Size. (GTM-9073)