[V63007] Fix mpt_extra/gbl and mpt_extra/gbl_utf8 subtest failures due to changes in %GCE, %GD, %GO, %GSE, %GSEL (GTM-8130)
Below is the release note for GTM-8130.

The %GSEL utility now silently ignores any subscript in the search string and throws a non-fatal error
if the input contains an invalid character. Previously, %GSEL would remove invalid characters and then
perform the search. This could cause problems if a subscript was present because the utility removed
"(" and ")" from the search pattern but not what was between them. This change also applies to %GCE,
%GD, %GO, and %GSE which use GD^%GSEL to search for globals. In addition, when used interactively,
it attempts to preserve the original I/O state of the caller; previously it tended to leave that
state disrupted. (GTM-8130)