[V63007] Fix socket/lsocpass and v63005/gtm8961 subtest failures due to socketpool device being renamed to YGTMSOCKETPOOL (GTM-4796)
Below is the release note of GTM-4796.

SOCKET devices use "YGTMSOCKETPOOL" to identify the socket pool; an attempt OPEN a device of that
name produces a DEVNAMERESERVED error. Note this change requires adjustment of any code explicitly
referencing the socket pool. Previously, SOCKET devices used the name "socketpool" to designate
the socket pool and an OPEN of a device with that name prevented the use of the pool or access to
any devices sockets in it. (GTM-4796)