[V63007] Fix spanning_regions/collationtests and spanning_regions/gde subtest failure due to GTM-9065
Because GDE now treats canonic numeric string subscripts as numeric subscripts,
there were a few places in collationtests.gde where a NAMRANGEOVERLAP error now
occurs. For example, $ZCHAR(50), formerly a string "1" subscript, is now the
numeric subscript 1. Which confuses GDE when used in a range specification (:).
This is now fixed by using $ZCHAR(47) and $ZCHAR(46) (which do not translate to
a numeric subscript) instead of $ZCHAR(50) and $ZCHAR(49) respectively.

Similar fixes were needed in the spanning_regions/gde subtest.